learning macros in Assembly
A macro procedure is a named block of assembly language statements. Once defined, it can be invoked 
(called)  as  many  times  in  a  program  as  you  wish.  When  you  invoke  a  macro  procedure,  a  copy  of  its 
statements  is  inserted  directly  into  the  program.  It  is  customary  to  refer  to  calling  a  macro  procedure, although technically there is no CALL instruction involved.