Str_compare Procedure in assembly
The  strings  are  compared  byte  by  byte,  using  their  8-bit  integer  ASCII  codes.  The  comparison  is  case-sensitive because ASCII codes are different for uppercase and lowercase letters. The procedure does not return a value, but the Carry and Zero flags can be interpreted. look at the code. and run it to see the difference